Refactor(optim) + Fixes
